refactor: Extracted TextAvatarPicker from AvatarPicker
